Initial Sidebar Fixes
